Change dom implementations to store values as DOM attributes rather than members.
git-svn-id: https://svn.lojjic.net/xul/trunk@566 8f8d9d7b-adf4-0310-9a85-ad1d9f542d00

+++ xul-dom-api/src/main/java/net/lojjic/xul/XULControlElement.java
@@ -5,8 +5,8 @@ public interface XULControlElement extends XULElement {
 	boolean isDisabled();
 	void setDisabled(boolean disabled);
 	
-	long getTabIndex();
-	void setTabIndex(long tabIndex);
+	int getTabIndex();
+	void setTabIndex(int tabIndex);
 	
 	// XXX defined in XULElement, but should be defined here
 	//  readonly attribute nsIControllers controllers;
